Add a test for creating a temporary queue with unsupported attributes

+++ src/test/java/com/amazonaws/services/sqs/AmazonSQSTemporaryQueuesClientIT.java
@@ -8,6 +8,8 @@ import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 
+import com.amazonaws.services.sqs.model.CreateQueueRequest;
+import com.amazonaws.services.sqs.model.QueueAttributeName;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
@@ -31,13 +33,15 @@ public class AmazonSQSTemporaryQueuesClientIT extends IntegrationTest {
     
     @After
     public void teardown() {
-        client.deleteQueue(queueUrl);
+        if (queueUrl != null) {
+            client.deleteQueue(queueUrl);
+        }
         client.shutdown();
     }
     
     @Test
     public void createQueueAddsAttributes() {
-        queueUrl = client.createQueue("TestQueue").getQueueUrl();
+        queueUrl = client.createQueue(queueNamePrefix + "TestQueue").getQueueUrl();
         Map<String, String> attributes = client.getQueueAttributes(queueUrl, Collections.singletonList("All")).getAttributes();
         String hostQueueUrl = attributes.get(VIRTUAL_QUEUE_HOST_QUEUE_ATTRIBUTE);
         assertNotNull(hostQueueUrl);
@@ -45,5 +49,17 @@ public class AmazonSQSTemporaryQueuesClientIT extends IntegrationTest {
         
         Map<String, String> hostQueueAttributes = client.getQueueAttributes(queueUrl, Collections.singletonList("All")).getAttributes();
         Assert.assertEquals("300", hostQueueAttributes.get(AmazonSQSIdleQueueDeletingClient.IDLE_QUEUE_RETENTION_PERIOD));
+    }
+
+    @Test
+    public void createQueueWithUnsupportedAttributes() {
+        try {
+            client.createQueue(new CreateQueueRequest()
+                    .withQueueName(queueNamePrefix + "InvalidQueue")
+                    .withAttributes(Collections.singletonMap(QueueAttributeName.FifoQueue.name(), "true")));
+            Assert.fail("Shouldn't be able to create a FIFO temporary queue");
+        } catch (IllegalArgumentException e) {
+            Assert.assertEquals("Cannot create a temporary queue with the following attributes: FifoQueue", e.getMessage());
+        }
     }
 }
